+++ frontend/src/store/reducers/authReducer.js
@@ -1,0 +1,1 @@
+//This is the auth reducer
